requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
asyncio==3.4.3
asyncpg==0.30.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
frozenlist==1.5.0
idna==3.10
loguru==0.7.3
magic-filter==1.0.12
modeltranslation==0.25
multidict==6.1.0
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.2.3
win32_setctime==1.2.0
yarl==1.18.3
